Interpreter is monadic; infinite Poxgrams bad!
We shall not have our language embedded in Haskell; instead, we shall
define a parser for Poxgrams and disallow recursion in the language, so
that it's not possible to create an infinite Poxgram. Then we will add a
Repeat term to the language, the only way in which infinite streams can
be created.